Deprecate Cast2
- Move implementation to CastAndValidate
- Add module documentation for CastAndValidate
